+++ scripts/sheets/location-sheet.js
@@ -53,9 +53,16 @@ export class LocationSheet extends CampaignCodexBaseSheet {
                     name: journal.name,
                     img: journal.img || "icons/svg/book.svg" 
                 };
-            }
+            } else {
+            throw new Error(`Journal document not found for UUID.`);
+        }
         } catch (error) {
             console.warn(`Campaign Codex | Linked standard journal not found: ${locationData.linkedStandardJournal}`);
+            if (game.user.isGM) {
+            const updatedData = foundry.utils.deepClone(locationData);
+            updatedData.linkedStandardJournal = null;
+            await this.document.setFlag("campaign-codex", "data", updatedData);
+        }
         }
     }