+++ scripts/sheets/npc-sheet.js
@@ -43,12 +43,18 @@ export class NPCSheet extends CampaignCodexBaseSheet {
                     name: journal.name,
                     img: journal.img || "icons/svg/book.svg" 
                 };
-            }
+            } else {
+            throw new Error(`Journal document not found for UUID.`);
+        }
         } catch (error) {
             console.warn(`Campaign Codex | Linked standard journal not found: ${npcData.linkedStandardJournal}`);
+            if (game.user.isGM) {
+            const updatedData = foundry.utils.deepClone(npcData);
+            updatedData.linkedStandardJournal = null;
+            await this.document.setFlag("campaign-codex", "data", updatedData);
+        }
         }
     }
-